+++ src/Compiler/IR/HIR/HIRBuilder.cs
@@ -345,18 +345,20 @@ public static class HIRBuilder
                                     }
 
                                     ctorStatements.Add(new HIRExpressionStatement(
-                                        new HIRIndexAssignmentExpression(new HIRThisExpression(), computedKeyExpr, Acornima.Operator.Assignment, initExpr!)));
+                                        new HIRDefineClassDataPropertyExpression(
+                                            new HIRThisExpression(),
+                                            computedKeyExpr,
+                                            initExpr!,
+                                            isField: true)));
                                 }
-                                else if (hasResolvedInstanceFieldName)
+                                else if (propertyDefinition.Computed)
                                 {
-                                    ctorStatements.Add(new HIRStoreUserClassInstanceFieldStatement
-                                    {
-                                        RegistryClassName = registryClassName,
-                                        FieldName = computedInstanceFieldName!,
-                                        IsPrivateField = false,
-                                        Value = initExpr!,
-                                        Location = SourceLocation.FromNode(propertyDefinition)
-                                    });
+                                    ctorStatements.Add(new HIRExpressionStatement(
+                                        new HIRDefineClassDataPropertyExpression(
+                                            new HIRThisExpression(),
+                                            new HIRLiteralExpression(JavascriptType.String, computedInstanceFieldName!),
+                                            initExpr!,
+                                            isField: true)));
                                 }
                                 else if (propertyDefinition.Key is PrivateIdentifier priv)
                                 {
@@ -512,18 +514,20 @@ public static class HIRBuilder
                                     }
 
                                     initStatements.Add(new HIRExpressionStatement(
-                                        new HIRIndexAssignmentExpression(new HIRThisExpression(), computedKeyExpr, Acornima.Operator.Assignment, initExpr!)));
+                                        new HIRDefineClassDataPropertyExpression(
+                                            new HIRThisExpression(),
+                                            computedKeyExpr,
+                                            initExpr!,
+                                            isField: true)));
                                 }
-                                else if (hasResolvedInstanceFieldName)
+                                else if (propertyDefinition.Computed)
                                 {
-                                    initStatements.Add(new HIRStoreUserClassInstanceFieldStatement
-                                    {
-                                        RegistryClassName = registryClassName,
-                                        FieldName = computedInstanceFieldName!,
-                                        IsPrivateField = false,
-                                        Value = initExpr!,
-                                        Location = SourceLocation.FromNode(propertyDefinition)
-                                    });
+                                    initStatements.Add(new HIRExpressionStatement(
+                                        new HIRDefineClassDataPropertyExpression(
+                                            new HIRThisExpression(),
+                                            new HIRLiteralExpression(JavascriptType.String, computedInstanceFieldName!),
+                                            initExpr!,
+                                            isField: true)));
                                 }
                                 else if (propertyDefinition.Key is PrivateIdentifier priv)
                                 {
@@ -1469,7 +1473,7 @@ class HIRMethodBuilder
                             && ClassElementNames.TryGetPropertyName(propertyDefinition.Key, computed: true, out computedStaticFieldName)
                             && !string.IsNullOrWhiteSpace(computedStaticFieldName);
 
-                        if (propertyDefinition.Computed && (!hasResolvedStaticFieldName || string.Equals(computedStaticFieldName, "prototype", StringComparison.Ordinal)))
+                        if (propertyDefinition.Computed)
                         {
                             HIRExpression hirKey;
                             if (hasResolvedStaticFieldName)
@@ -1489,7 +1493,11 @@ class HIRMethodBuilder
                             }
 
                             statements.Add(new HIRExpressionStatement(
-                                new HIRIndexAssignmentExpression(classTypeExpr, hirKey, Acornima.Operator.Assignment, hirValue!)));
+                                new HIRDefineClassDataPropertyExpression(
+                                    classTypeExpr,
+                                    hirKey,
+                                    hirValue!,
+                                    isField: true)));
                             break;
                         }
 
@@ -3585,14 +3593,14 @@ class HIRMethodBuilder
                         if (getterMethodName != null)
                         {
                             hirExpr = new HIRCallExpression(
-                                new HIRPropertyAccessExpression(new HIRThisExpression(_staticThisRegistryClassName), getterMethodName),
+                                new HIRPropertyAccessExpression(new HIRThisExpression(), getterMethodName),
                                 Array.Empty<HIRExpression>());
                             return true;
                         }
 
                         if (privateMethodName != null)
                         {
-                            hirExpr = new HIRPropertyAccessExpression(new HIRThisExpression(_staticThisRegistryClassName), privateMethodName);
+                            hirExpr = new HIRPropertyAccessExpression(new HIRThisExpression(), privateMethodName);
                             return true;
                         }